made that everynew user get a free sub by default, created the change subscription modal, started working on the context to transfer data subscription from grandchildern to grandparent component and then pass it as a prop to the modal that in turn will make the post request to the db

+++ models/subscription.js
@@ -2,11 +2,15 @@ module.exports = (sequelize, DataTypes) => {
   const Subscription = sequelize.define('Subscription', {
     name: { type: DataTypes.STRING, allowNULL: false, defaultValue: 'Green' },
     price: {
-      type: DataTypes.INT,
+      type: DataTypes.FLOAT,
       allowNULL: false,
       defaultValue: 0,
     },
-    freeListings: { type: DataTypes.INT, allowNULL: false, defaultValue: 10 },
+    freeListings: {
+      type: DataTypes.INTEGER,
+      allowNULL: false,
+      defaultValue: 10,
+    },
     pricePerListings: {
       type: DataTypes.FLOAT,
       allowNULL: false,
